# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "JalNetra",
  "description": "Satellite-based water quality anomaly intelligence for Maharashtra.\n\n**Boundary:** the system detects *potential anomalies* in satellite-observable indicators and prioritises zones for ground investigation. It never claims to detect, prove or confirm pollution or contamination.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/tiles/chip/{chip}/{z}/{x}/{y}.png": {
   "get": {
    "tags": [
     "tiles"
    ],
    "summary": "Chip Tile",
    "operationId": "chip_tile_tiles_chip__chip___z___x___y__png_get",
    "parameters": [
     {
      "name": "chip",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "description": "URL-encoded MinIO object key of a COG chip",
       "title": "Chip"
      },
      "description": "URL-encoded MinIO object key of a COG chip"
     },
     {
      "name": "z",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Z"
      }
     },
     {
      "name": "x",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "X"
      }
     },
     {
      "name": "y",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Y"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/tiles/{layer}/{water_body_id}/{on}/{z}/{x}/{y}.png": {
   "get": {
    "tags": [
     "tiles"
    ],
    "summary": "Layer Tile",
    "description": "Body-scope chip of ``layer`` (``watermask``, ``ndti_turbidity``, ... or an alias\nsuch as ``turbidity``) for a water body on a date.",
    "operationId": "layer_tile_tiles__layer___water_body_id___on___z___x___y__png_get",
    "parameters": [
     {
      "name": "layer",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Layer"
      }
     },
     {
      "name": "water_body_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Water Body Id"
      }
     },
     {
      "name": "on",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "date",
       "title": "On"
      }
     },
     {
      "name": "z",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Z"
      }
     },
     {
      "name": "x",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "X"
      }
     },
     {
      "name": "y",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Y"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/validations": {
   "post": {
    "tags": [
     "validations"
    ],
    "summary": "Create Validation",
    "description": "Submit a field or lab result against an alert. The verdict (matched /\nnot_matched / inconclusive) is computed by the validation loop (S11).",
    "operationId": "create_validation_api_v1_validations_post",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ValidationIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "validations"
    ],
    "summary": "List Validations",
    "description": "Submitted validations, newest first, with their verdict when computed.",
    "operationId": "list_validations_api_v1_validations_get",
    "parameters": [
     {
      "name": "alert_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Alert Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationList"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/jobs/ingest": {
   "post": {
    "tags": [
     "jobs"
    ],
    "summary": "Create Ingest Job",
    "description": "Enqueue ingestion for a water body and date range. Each ingested scene\nchains mask -> indicators -> anomalies -> scoring -> alerts on its own;\npoll ``GET /jobs/{id}`` for the live readout.",
    "operationId": "create_ingest_job_api_v1_jobs_ingest_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/IngestJobRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/jobs": {
   "get": {
    "tags": [
     "jobs"
    ],
    "summary": "List Jobs",
    "operationId": "list_jobs_api_v1_jobs_get",
    "parameters": [
     {
      "name": "water_body_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Water Body Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobList"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/jobs/{job_id}": {
   "get": {
    "tags": [
     "jobs"
    ],
    "summary": "Get Job",
    "description": "State, progress percent and the stage currently in flight.",
    "operationId": "get_job_api_v1_jobs__job_id__get",
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Job Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "IngestJobRequest": {
    "properties": {
     "water_body_id": {
      "type": "string",
      "title": "Water Body Id"
     },
     "date_from": {
      "type": "string",
      "format": "date",
      "title": "Date From"
     },
     "date_to": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date To",
      "description": "defaults to date_from"
     },
     "requested_by": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Requested By"
     }
    },
    "type": "object",
    "required": [
     "water_body_id",
     "date_from"
    ],
    "title": "IngestJobRequest",
    "examples": [
     {
      "date_from": "2026-09-01",
      "date_to": "2026-09-21",
      "water_body_id": "wb_khadakwasla"
     }
    ]
   },
   "JobList": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/JobOut"
      },
      "type": "array",
      "title": "Items"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     }
    },
    "type": "object",
    "required": [
     "items",
     "next_cursor"
    ],
    "title": "JobList"
   },
   "JobOut": {
    "properties": {
     "job_id": {
      "type": "string",
      "title": "Job Id"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "water_body_id": {
      "type": "string",
      "title": "Water Body Id"
     },
     "date_from": {
      "type": "string",
      "format": "date",
      "title": "Date From"
     },
     "date_to": {
      "type": "string",
      "format": "date",
      "title": "Date To"
     },
     "status": {
      "type": "string",
      "enum": [
       "queued",
       "running",
       "done",
       "failed"
      ],
      "title": "Status"
     },
     "progress_pct": {
      "type": "number",
      "maximum": 100.0,
      "minimum": 0.0,
      "title": "Progress Pct"
     },
     "current_stage": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Current Stage",
      "description": "first stage with pending work; null when done"
     },
     "stages": {
      "items": {
       "$ref": "#/components/schemas/StageProgress"
      },
      "type": "array",
      "title": "Stages"
     },
     "scenes_found": {
      "type": "integer",
      "title": "Scenes Found"
     },
     "scenes_usable": {
      "type": "integer",
      "title": "Scenes Usable"
     },
     "alerts_created": {
      "type": "integer",
      "title": "Alerts Created"
     },
     "celery_state": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Celery State"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     },
     "finished_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     }
    },
    "type": "object",
    "required": [
     "job_id",
     "kind",
     "water_body_id",
     "date_from",
     "date_to",
     "status",
     "progress_pct",
     "current_stage",
     "stages",
     "scenes_found",
     "scenes_usable",
     "alerts_created",
     "celery_state",
     "error",
     "created_at",
     "updated_at",
     "finished_at"
    ],
    "title": "JobOut",
    "examples": [
     {
      "alerts_created": 1,
      "celery_state": "SUCCESS",
      "created_at": "2026-09-21T10:00:00Z",
      "current_stage": "indicators",
      "date_from": "2026-09-01",
      "date_to": "2026-09-21",
      "job_id": "job_3f9c2b7e4a5d4c1e9b8a7f6e5d4c3b2a",
      "kind": "ingest",
      "progress_pct": 58.3,
      "scenes_found": 5,
      "scenes_usable": 4,
      "stages": [
       {
        "done": 4,
        "pct": 100,
        "stage": "ingestion",
        "total": 4
       },
       {
        "done": 4,
        "pct": 100,
        "stage": "mask",
        "total": 4
       },
       {
        "done": 2,
        "pct": 50,
        "stage": "indicators",
        "total": 4
       },
       {
        "done": 1,
        "pct": 25,
        "stage": "anomalies",
        "total": 4
       },
       {
        "done": 1,
        "pct": 25,
        "stage": "scoring",
        "total": 4
       },
       {
        "done": 1,
        "pct": 25,
        "stage": "alerts",
        "total": 4
       }
      ],
      "status": "running",
      "updated_at": "2026-09-21T10:03:12Z",
      "water_body_id": "wb_khadakwasla"
     }
    ]
   },
   "LabResults": {
    "properties": {
     "turbidity_ntu": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Turbidity Ntu"
     },
     "chlorophyll_ug_l": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Chlorophyll Ug L"
     },
     "tss_mg_l": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Tss Mg L"
     },
     "do_mg_l": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Do Mg L"
     },
     "ph": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 14.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Ph"
     },
     "temperature_c": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Temperature C"
     },
     "conductivity_us_cm": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Conductivity Us Cm"
     }
    },
    "type": "object",
    "title": "LabResults"
   },
...
   "StageProgress": {
    "properties": {
     "stage": {
      "type": "string",
      "title": "Stage"
     },
     "done": {
      "type": "integer",
      "title": "Done"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "pct": {
      "type": "number",
      "title": "Pct"
     }
    },
    "type": "object",
    "required": [
     "stage",
     "done",
     "total",
     "pct"
    ],
    "title": "StageProgress"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "ValidationIn": {
    "properties": {
     "alert_id": {
      "type": "string",
      "title": "Alert Id"
     },
     "sampled_on": {
      "type": "string",
      "format": "date",
      "title": "Sampled On"
     },
     "lab_results": {
      "$ref": "#/components/schemas/LabResults"
     },
     "observed_condition": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 2000
       },
       {
        "type": "null"
       }
      ],
      "title": "Observed Condition"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 4000
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "submitted_by": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Submitted By"
     }
    },
    "type": "object",
    "required": [
     "alert_id",
     "sampled_on"
    ],
    "title": "ValidationIn",
    "examples": [
     {
      "alert_id": "alr_2026_0917_khadakwasla_z3",
      "lab_results": {
       "ph": 7.6,
       "tss_mg_l": 62.0,
       "turbidity_ntu": 48.0
      },
      "observed_condition": "Brown plume near the eastern inlet, no odour.",
      "sampled_on": "2026-09-19",
      "submitted_by": "RO Pune field team"
     }
    ]
   },
   "ValidationList": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ValidationOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "next_cursor"
    ],
    "title": "ValidationList"
   },
   "ValidationOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "alert_id": {
      "type": "string",
      "title": "Alert Id"
     },
     "sampled_on": {
      "type": "string",
      "format": "date",
      "title": "Sampled On"
     },
     "lab_results": {
      "additionalProperties": true,
      "type": "object",
      "title": "Lab Results"
     },
     "observed_condition": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Observed Condition"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "submitted_by": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Submitted By"
     },
     "verdict": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "matched",
         "not_matched",
         "inconclusive"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Verdict"
     },
     "verdict_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Verdict Reason"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "alert_id",
     "sampled_on",
     "lab_results",
     "observed_condition",
     "notes",
     "submitted_by",
     "verdict",
     "verdict_reason",
     "created_at"
    ],
    "title": "ValidationOut"
   },
...
  }
 }
}